requirements.txt
flask
python-dotenv
google-genai
Pillow